+++ lib/malloy/client/type_traits.hpp
@@ -5,39 +5,13 @@
 #include "../core/type_traits.hpp"
 
 #include "malloy/core/tmp.hpp"
+#include "malloy/client/tmp.hpp"
 
 #include <concepts>
 #include <variant>
 
 
-namespace malloy::client {
-namespace tmp {
-
-namespace detail {
-template<template<typename...> typename V, typename... Vargs>
-struct conv_to_resp_helper {
-    constexpr conv_to_resp_helper(V<Vargs...>) {}
-    using type = V<http::response<Vargs>...>;
-};
-}
-
-template<typename V>
-requires (malloy::concepts::is_variant_of_bodies<V>)
-using body_type = malloy::tmp::unwrap_variant<V>;
-
-template<typename F>
-using bodies_for_t = std::invoke_result_t<decltype(&F::body_for), const F*, const typename F::header_type&>;
-
-template<typename Bodies>
-requires (malloy::concepts::is_variant_of_bodies<Bodies>)
-using to_responses = typename decltype(detail::conv_to_resp_helper{std::declval<Bodies>()})::type;
-
-template<typename Filter>
-using filter_resp_t = malloy::tmp::unwrap_variant<to_responses<bodies_for_t<Filter>>>;
-
-};
-
-namespace concepts
+namespace malloy::client::concepts
 {
 
     namespace detail
@@ -97,7 +71,6 @@ namespace concepts
         std::visit(detail::http_cb_helper<F>{std::move(cb)}, f.body_for(h));
     };
 
-}
 }
 
 /**
